requirements.txt
python-telegram-bot==21.5
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pyyaml==6.0.1
ccxt==4.2.18
ta==0.10.2
pytest==7.4.3
schedule==1.2.0
aiohttp==3.9.1